requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4; python_version >= '3'
cryptography==3.4.7
defusedxml==0.7.1
deprecated==1.2.12
django-autocomplete-light==3.8.2
django-bootstrap-breadcrumbs==0.9.2
django-braces==1.14.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-oauth-toolkit==1.5.0
django-phone-field==1.8.1
django-redis==5.0.0
django-rest-framework-social-oauth2==1.1.0
django-tables2-column-shifter==2.0.3
django-tables2==2.4.0
django-tempus-dominus==5.1.2.14
django-widget-tweaks==1.4.8
django==3.2.6
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2; python_version >= '3'
jwcrypto==1.0
oauthlib==3.1.1
psycopg2-binary==2.9.1
pycparser==2.20
pyjwt==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests-oauthlib==1.3.0
requests==2.26.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.6
wrapt==1.12.1
